requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0

# MLflow
mlflow>=2.5.0

# Logging
structlog>=23.0.0

# Path handling
pathlib2>=2.3.7; python_version < '3.4'

